Diagrams.Backend.Postscript.CMYK

Description

Support for CMYK color attributes in the Postscript backend.

CMYK

CMYK colors are represented with four values from 0.0 to 1.0.

data CMYK

Constructors

CMYK

Fields

_cyan :: Double

_magenta :: Double

_yellow :: Double

_blacK :: Double

Line color

data LineColorCMYK

The color with which lines (strokes) are drawn. Note that child colors always override parent colors; that is, lineColorCMYK c1 . lineColorCMYK c2 $ d is equivalent to lineColorCMYK c2 $ d. More precisely, the semigroup structure on line color attributes is that of Last.

getLineColorCMYK :: LineColorCMYK -> CMYK

mkLineColorCMYK :: CMYK -> LineColorCMYK

styleLineColorCMYK :: Setter' (Style v) CMYK

lineColorCMYK :: HasStyle a => CMYK -> a -> a

Set the line (stroke) color.

lineColorCMYKA :: HasStyle a => LineColorCMYK -> a -> a

Apply a lineColorCMYK attribute.

lcCMYK :: HasStyle a => CMYK -> a -> a

A synonym for lineColorCMYK.

Fill color

data FillColorCMYK

The color with which shapes are filled. Note that child colors always override parent colors; that is, fillColorCMYK c1 . fillColorCMYK c2 $ d is equivalent to lineColorCMYK c2 $ d. More precisely, the semigroup structure on fill color attributes is that of Last.

getFillColorCMYK :: FillColorCMYK -> CMYK

mkFillColorCMYK :: CMYK -> FillColorCMYK

styleFillColorCMYK :: Setter' (Style v) CMYK

recommendFillColorCMYK :: HasStyle a => CMYK -> a -> a

Set a "recommended" fill color, to be used only if no explicit calls to fillColor (or fc, or fcA) are used.

fillColorCMYK :: HasStyle a => CMYK -> a -> a

Set the fill color.

fcCMYK :: HasStyle a => CMYK -> a -> a

A synonym for fillColorCMYK