Diagrams.Backend.Postscript.CmdLine

Description

Convenient creation of command-line-driven executables for rendering diagrams using the Postscript backend.

Documentation

defaultMain :: Diagram Postscript R2 -> IO ()

multiMain :: [(String, Diagram Postscript R2)] -> IO ()

pagesMain :: [Diagram Postscript R2] -> IO ()

data Postscript

This data declaration is simply used as a token to distinguish this rendering engine.