Diagrams.Backend.Postscript

Description

The EPS backend.

Documentation

data Postscript

This data declaration is simply used as a token to distinguish this rendering engine.

Constructors

Postscript

data family Options b v

Backend-specific rendering options.

data OutputFormat

Postscript could output to several file formats, which each have their own associated properties that affect the output.

Constructors

EPS