Diagrams.Backend.Postscript

Description

A Postscript rendering backend for diagrams.

To build diagrams for Postscript rendering use the Postscript type in the diagram type construction

d :: Diagram Postscript
d = ...

and render giving the Postscript token

renderDia Postscript (PostscriptOptions "file.eps" (Width 400) EPS) d

This IO action will write the specified file.

Backend token

data Postscript

This data declaration is simply used as a token to distinguish this rendering engine.

Constructors

Postscript

type B = Postscript

Postscript-specific options

Unfortunately, Haddock does not yet support documentation for associated data families, so we must just provide it manually. This module defines

data family Options Postscript V2 Double = PostscriptOptions
          { _psfileName     :: String             -- ^ the name of the file you want generated
          , _psSizeSpec     :: SizeSpec V2 Double -- ^ the requested size of the output
          , _psOutputFormat :: OutputFormat        -- ^ the output format and associated options
          }

data family Options b (v :: * -> *) n :: *

Backend-specific rendering options.

psfileName :: Lens' (Options Postscript V2 Double) String

psSizeSpec :: Lens' (Options Postscript V2 Double) (SizeSpec V2 Double)

psOutputFormat :: Lens' (Options Postscript V2 Double) OutputFormat

Postscript-supported output formats

data OutputFormat

Postscript only supports EPS style output at the moment. Future formats would each have their own associated properties that affect the output.

Constructors

EPS
Encapsulated Postscript output.

renderDias :: (Semigroup m, Monoid m) => Options Postscript V2 Double -> [QDiagram Postscript V2 Double m] -> IO [()]